config/requirements/base-requirements.txt
django==4.1.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
celery==5.2.7
redis==4.3.4
django-redis==5.2.0
requests==2.28.1
requests-mock==1.10.0
